add loading stores and loading bar
